+++ CHANGELOG.md
@@ -8,7 +8,7 @@ In Development
 - Format all log lines as JSON
 - Add logging of Zarr manifest cache events
 - Limit Zarr manifest cache by total size of entries
-    - Add a `-Z`/`--zarrman-cache-bytes` option for setting the cache size
+    - Add a `-Z`/`--zarrman-cache-mb` option for setting the cache size
 - Expire idle Zarr manifest cache entries
 - Log Zarr manifest cache entries every hour